+++ packages/cli/src/commands/dev/dev.js
@@ -1,5 +1,5 @@
 /*
-  Copyright 2020-2021 Lowdefy, Inc
+  Copyright 2020-2022 Lowdefy, Inc
 
   Licensed under the Apache License, Version 2.0 (the "License");
   you may not use this file except in compliance with the License.
@@ -15,6 +15,7 @@
 */
 
 import addCustomPluginsAsDeps from '../../utils/addCustomPluginsAsDeps.js';
+import copyPluginsFolder from '../../utils/copyPluginsFolder.js';
 import installServer from '../../utils/installServer.js';
 import runDevServer from './runDevServer.js';
 import getServer from '../../utils/getServer.js';
@@ -23,6 +24,7 @@ async function dev({ context }) {
   const directory = context.directories.dev;
   context.print.info('Starting development server.');
   await getServer({ context, packageName: '@lowdefy/server-dev', directory });
+  await copyPluginsFolder({ context, directory });
   await addCustomPluginsAsDeps({ context, directory });
   await installServer({ context, directory });
   context.sendTelemetry();