+++ packages/operators/src/nodeParser.js
@@ -1,5 +1,5 @@
 /*
-  Copyright 2020-2021 Lowdefy, Inc
+  Copyright 2020-2022 Lowdefy, Inc
 
   Licensed under the Apache License, Version 2.0 (the "License");
   you may not use this file except in compliance with the License.
@@ -17,7 +17,8 @@
 import { serializer, type } from '@lowdefy/helpers';
 
 class NodeParser {
-  constructor({ payload, secrets, user, operators }) {
+  constructor({ env, payload, secrets, user, operators }) {
+    this.env = env;
     this.operators = operators;
     this.payload = payload;
     this.secrets = secrets;
@@ -25,7 +26,8 @@ class NodeParser {
     this.parse = this.parse.bind(this);
   }
 
-  parse({ args, input, location }) {
+  parse({ args, input, location, operatorPrefix = '_' }) {
+    const env = this.env;
     const operators = this.operators;
     const secrets = this.secrets;
     const payload = this.payload;
@@ -42,33 +44,34 @@ class NodeParser {
     }
     const errors = [];
     const reviver = (_, value) => {
-      if (type.isObject(value) && Object.keys(value).length === 1) {
-        const key = Object.keys(value)[0];
-        const [op, methodName] = key.split('.');
-        try {
-          if (!type.isUndefined(operators[op])) {
-            const res = operators[op]({
-              args,
-              arrayIndices: [],
-              env: 'node',
-              location,
-              methodName,
-              operators: operators,
-              params: value[key],
-              secrets,
-              payload,
-              user,
-              parser: this,
-            });
-            return res;
-          }
-        } catch (e) {
-          errors.push(e);
-          console.error(e);
-          return null;
-        }
+      if (!type.isObject(value) || Object.keys(value).length !== 1) return value;
+
+      const key = Object.keys(value)[0];
+      if (!key.startsWith(operatorPrefix)) return value;
+
+      const [op, methodName] = `_${key.substring(operatorPrefix.length)}`.split('.');
+      if (type.isUndefined(operators[op])) return value;
+      try {
+        const res = operators[op]({
+          args,
+          arrayIndices: [],
+          env,
+          location,
+          methodName,
+          operators: operators,
+          params: value[key],
+          operatorPrefix,
+          parser: this,
+          payload,
+          secrets,
+          user,
+        });
+        return res;
+      } catch (e) {
+        errors.push(e);
+        console.error(e);
+        return null;
       }
-      return value;
     };
     return {
       output: serializer.copy(input, { reviver }),