+++ packages/plugins/operators/operators-js/src/operators/shared/array.test.js
@@ -1,5 +1,5 @@
 /*
-  Copyright 2020-2021 Lowdefy, Inc
+  Copyright 2020-2022 Lowdefy, Inc
 
   Licensed under the Apache License, Version 2.0 (the "License");
   you may not use this file except in compliance with the License.
@@ -31,6 +31,8 @@ const operators = {
 };
 
 const location = 'location';
+
+const operatorPrefix = '_';
 
 const parser = new NodeParser({ operators, payload: {}, secrets: {}, user: {} });
 
@@ -151,7 +153,12 @@ describe('_array.copyWithin', () => {
 
 describe('_array.every', () => {
   const methodName = 'every';
-  const callback = _function({ location, params: { __gt: [{ __args: '0' }, 3] }, parser });
+  const callback = _function({
+    location,
+    operatorPrefix,
+    params: { __gt: [{ __args: '0' }, 3] },
+    parser,
+  });
   test('valid', () => {
     expect(
       _array({
@@ -272,7 +279,12 @@ describe('_array.fill', () => {
 
 describe('_array.filter', () => {
   const methodName = 'filter';
-  const callback = _function({ location, params: { __gt: [{ __args: '0' }, 3] }, parser });
+  const callback = _function({
+    location,
+    operatorPrefix,
+    params: { __gt: [{ __args: '0' }, 3] },
+    parser,
+  });
   test('valid', () => {
     expect(
       _array({
@@ -327,7 +339,12 @@ describe('_array.filter', () => {
 
 describe('_array.find', () => {
   const methodName = 'find';
-  const callback = _function({ location, params: { __gt: [{ __args: '0' }, 3] }, parser });
+  const callback = _function({
+    location,
+    operatorPrefix,
+    params: { __gt: [{ __args: '0' }, 3] },
+    parser,
+  });
   test('valid', () => {
     expect(
       _array({
@@ -382,7 +399,12 @@ describe('_array.find', () => {
 
 describe('_array.findIndex', () => {
   const methodName = 'findIndex';
-  const callback = _function({ location, params: { __gt: [{ __args: '0' }, 3] }, parser });
+  const callback = _function({
+    location,
+    operatorPrefix,
+    params: { __gt: [{ __args: '0' }, 3] },
+    parser,
+  });
   test('valid', () => {
     expect(
       _array({
@@ -703,7 +725,12 @@ describe('_array.lastIndexOf', () => {
 
 describe('_array.map', () => {
   const methodName = 'map';
-  const callback = _function({ location, params: { __sum: [{ __args: '0' }, 1] }, parser });
+  const callback = _function({
+    location,
+    operatorPrefix,
+    params: { __sum: [{ __args: '0' }, 1] },
+    parser,
+  });
   test('valid', () => {
     expect(
       _array({
@@ -760,6 +787,7 @@ describe('_array.reduce', () => {
   const methodName = 'reduce';
   const callback = _function({
     location,
+    operatorPrefix,
     params: { __sum: [{ __args: '0' }, { __args: '1' }] },
     parser,
   });
@@ -830,6 +858,7 @@ describe('_array.reduceRight', () => {
   const methodName = 'reduceRight';
   const callback = _function({
     location,
+    operatorPrefix,
     params: { __sum: [{ __args: '0' }, { __args: '1' }] },
     parser,
   });
@@ -1044,7 +1073,12 @@ describe('_array.splice', () => {
 
 describe('_array.some', () => {
   const methodName = 'some';
-  const callback = _function({ location, params: { __gt: [{ __args: '0' }, 3] }, parser });
+  const callback = _function({
+    location,
+    operatorPrefix,
+    params: { __gt: [{ __args: '0' }, 3] },
+    parser,
+  });
   test('valid', () => {
     expect(
       _array({