+++ packages/utils/block-utils/src/mediaToCssObject.js
@@ -1,6 +1,6 @@
 /* eslint-disable no-undef */
 /*
-  Copyright 2020-2021 Lowdefy, Inc
+  Copyright 2020-2022 Lowdefy, Inc
 
   Licensed under the Apache License, Version 2.0 (the "License");
   you may not use this file except in compliance with the License.